changed uint size in Constants contract

+++ CurrentBest.sol
@@ -13,14 +13,13 @@ contract GasContract is Ownable, Constants {
     uint256 public totalSupply = 0; // cannot be updated
     uint256 public paymentCounter = 0;
     mapping(address => uint256) public balances;
-    uint8 public tradePercent = 12;
+    uint256 public tradePercent = 12;
     address public contractOwner;
     uint256 public tradeMode = 0;
     mapping(address => Payment[]) public payments;
     mapping(address => uint256) public whitelist;
     address[5] public administrators;
     bool public isReady = false;
-    
     enum PaymentType {
         Unknown,
         BasicPayment,
